// wedding-front/projects/mariage/src/app/modules/user/components/company/company-details/category-label.enum.ts
export enum CategoryLabelEnum {
    PHOTOGRAPHE = 'photographe',
    VIDEALISTE = 'videaliste',
    DECORATUER = 'decorateur',
    FLEURISTE = 'fleuriste',
    VOITURE = 'voiture',
    BUS = 'bus',
    LIEU = 'lieu',
    TRAITEUR = 'traiteur',
    GATEAU_MARIAGE = 'gateaumariage',
    DJ = 'dj',
    GROUPE = 'groupe',
    MUSICIEN = 'musicien',
    ANIMATEUR_ADULTS = 'animateurAdultes',
    ANIMATEUR_ENFANTS = 'animateurEnfants',
    FEU_ARTIFICES = 'feuArtifices',
    LACHERS = 'lacher',
    COACH = 'coachSportif',
    CHOREGRAPHE = 'choregrapheMariage',
    OFFICIANT = 'officiantCeremonie',
    COIFFURE = 'coiffure',
    MAQUILLAGE = 'maquillage',
    ESTHETIQUE = 'esthetique',
    SOIN = 'soins',
    HEBERGEMENT = 'hebergement',
    FAIRE_PART = 'faireparts',
    VOYAGE_DE_NOCES = 'voyagenoces',
}